sem5/OIIS/lab5/pyproject.toml
[tool.poetry]
name = "lab5"
version = "0.1.0"
description = ""
authors = ["Ehor Hleoza <gleoza04@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
